fixing issue with debug (popup mode)
$xoopslogger->dump()

if we are making function dump() in logger.php a protected function then we need to either create a public function inside the class that refers to the protection function, or we make the protected function dump() public.

otherwise we get errors Fatal error: Call to protected method XoopsLogger::dump() from context '' in /home/g7poo/public_html/alphatest/include/functions.php on line 690

why? because if you create a protected function in a class, then that function is obviously protected, which means that function can only be called from within that same class itself, and not from any other external class or function.

there's still more issues with popup debug.. ie. the data is displayed briefly in the popup for about a millisecond and then disappears, so the logger render is faulty too. which is making debugging all the more difficult than it already is.