#624: First, Phoenyx, you are right, the problem was still present on fresh install. I have edited makedata.php to fix this. Now, it's working. Can you test again a fresh install ?
And following your's and Vaughan's comments, yes, we need to improve this in the future. But considering the status of this release cycle, this is the only fix I can think of that has the minimum impact on code. Yes, we will loose the WYSIWYG editor on those preferences control, but I think it's a small price to pay to make it work for now. And we can improve in 1.3 or 1.4.

ref[t:624]