Addressing the issue described in Ticket #438.
One line bug fix.  The makeclickable method in the text sanitizer class was not always removing the preceding space that it added to text passed in.  This led to certain situations where text would end up with white space padding at the beginning after going through the text sanitizer.  Depending on the semantics of the data being passed in and out, and what the userland code was doing with it, this extra space could lead to major errors in logic.