Replacing appropriate instances of xoops_pagetitle throughout the core.
This is a partial solution - xoops_pagetitle and most other xoops_* smarty variables are being populated with the icms_* vars before the page is rendered. However, this does not account for any modules that may still make use of them, because we have not communicated this.

This commit only begins to address several places in the core where smarty variables were being populated but not being passed to the rendering of the page.