adding 2 new config preferences to install for HTMLPurifier 4.2.0
i can't remember how we add these in the update, but would be preferable to add them in the same order as in makedata.php. if someone can take care of it. new fields are;

	$dbm->insert('config', " VALUES (" . ++$i . ", 0, $c, 'purifier_HTML_FlashAllowFullScreen', '_MD_AM_PURIFIER_HTML_FLASHFULLSCRN', '0', '_MD_AM_PURIFIER_HTML_FLASHFULLSCRNDSC', 'yesno', 'int', " . $p++ . ")");

&

	$dbm->insert('config', " VALUES (" . ++$i . ", 0, $c, 'purifier_Core_NormalizeNewlines', '_MD_AM_PURIFIER_CORE_NORMALNEWLINES', '1', '_MD_AM_PURIFIER_CORE_NORMALNEWLINESDSC', 'yesno', 'int', " . $p++ . ")");

TODO: need to change references to icms_HTMLPurifier.php in core to use core/HTMLFilter.php or even just to use icms_core_DataFilter::checkVar($html, 'html'), think these are in the displayTarea() & previewTarea() functions in textSanitizer.