Applying code standards; replacing text with language constants; eliminating a few loops, using implode() instead
There were a LOT of changes, just from replacing text with language constants. One of the things I applied was removing as much markup from the constants as possible - the HTML tags shouldn't be in the content anyway. Along with removing the HTML tags, I also consolidated several duplicate constants and used a standard form of _MD_AM (for 'module' - 'admin') followed by a object, a method, and optionally a result. For example. _MD_AM_CONFIGOPTION_DELETE_FAIL and _MD_AM_CONFIGOPTION_DELETED. I think we should apply this wherever we can. It will make the translators' jobs a bit bigger at first, but it will help us eliminate multiple constants for the same thing.

There are still a few that need to be moved to a different language file - _MD_AM_SQL* is a more generic message and doesn't belong in a specific category of the admin area. I also did not spend a lot of time in install/modulesadmin.php - this file should be removed completely and the core file used instead.