Adding 2 required methods to the handler - they are abstract in the parent class.
We also need to examine this - there are classes in libraries/icms/data/rank and modules/system/admin/userrank/class - we only need them in one place. The system module really only provides an interface for managing the core object and should not have any objects of its own. This may be a large task to add to 1.3, but we need to prepare ourselves for this.