using datafilter filters.
working fine, except when using tinymce, bbcode doesn't work which is fine by me, bbcode isn't necessary when using a wysiwyg editor.

when using dhtmleditor and html is enabled, editing signature, converts bbcode to html ok. but doesn't insert a break. as i'm filtering on input, its an issue when using dhtmleditor for inputting html unless the user physically type <br /> instead of relying on nl2br, i think if ur using html that is acceptable. but users may get confused & wonder y there's no breaks.

are we getting rid of dhtmleditor in 1.3 or not? if we are that's superb. if not, it might be a lot better to restrict dhtmleditor to plaintext only, and force users to use wysiwyg if they want to use html.