fixing issue with URL tags when using TinyMCE editor on signatures.
problem is that function one_wordwrap does not take into consideration URL strings, and inserts a space in the string if it is longer than allowed.

not a problem in some circumstances, but as i have changed the method in edituser (the core edituser not profile module) so that it filters the user input before it gets to the DB, then smileys & xoopscode tags are converted to html before they are sent to the DB.

this means that when viewing userinfo.php the function one_wordwrap is breaking the url resulting in urls & images no displaying in the signature.

the function one_wordwrap & icms_substr oughta be modified to IGNORE url strings.