adding filter for Youtube new Iframe method. the new filter creates a span from the iframe code before filtering, then creates an <object> after filtering.
if we want to keep the iframe method for youtube videos, we'll need a filter for output in DataFilter to convert the youtube <object> to <iframe>

time for some experimentation, i may just have thought of a way of doing the filters a bit differently for use with our core.
also need to fix the custom.Filter preference, there's an error due to the serialization of the array in HTMLFilter.php tried unserialize.
might try & just create a multiselect box instead which reads the filenames of filters in the Filter folder to fill up the select options.