Reverting other changes where icms::$xoopsDB had been substituted for $this->db or even $db, where $db was created by icms_db_Factory::getInstance.
As it was explained in an email exchange:

In our current code, it is proper to replace all occurrences of $xoopsDB-> with icms::$xoopsDB-> and then remove the global $xoopsDB declaration from the files. But, in places were you have $this->db or $this->_db, the class has already made an assignment to the db object and no changes are necessary for each of those occurrences. To do so would only force more changes when switching to a different db engine/class. The change would only be at the point the assignment is made.