using Datafilter functions, replacing deprecated textsanitizer use.
note: it's not actually advisable to change some of the displayTarea() functions to use the checkVar() function in datafilter at this time. this is due to a necessary change required in doing so for HTML content, as we would need to make sure & convert the data already stored in the DB for those types to be filtered before we use it on output.

it is ok to change displayTarea($data, 0) to use DataFilter::checkVar($data, 'text', 'input/output'). as if displayTarea parameter for html is '0' then it's fine just make sure you use the right parameter for 'input' or output depending on it's use. leave displayTarea($data, 1) methods (HTML) as they are till 1.4.