Adding a new method to icms_core_DataFilter - checkVarArray. This allows you to easily create an array of input items and a filter array with options, passing them all to checkVar().
Also, checkVar always sets the options before calling priv_checkVar and using isset($option1) or isset($option2) will always return TRUE, even if the value is an empty string. This was causing the email check to always check against the bad email list and stopspammer. Made a change in priv_checkVar to allow the email validation to happen as requested.