refs #655 - OpenID not working.
The response back from the provider was not being stored properly in the session - __PHP_Incomplete_Class was how it was being stored. Serializing the object before saving it in the session seems to fix this, but I'm not sure this is the best approach. This usually happens when the class is not defined before the session starts - it may be a side effect of how the session and classes are being autoloaded.