Change in filtering for page_url - refs #712.
I was able to verify this works properly by first searching for o'niel (system search) If not properly sanitized, there was an error thrown by Protector. After the changes, the error is removed and the string is properly sanitized. 

There still is work to do on the symlinks feature, though. You can create a symlink that contains a single quote and it is not escaped or encoded when being inserted to the database. This means the new sanitized request URI will not match the stored symlink. (also verified - create a new symlink, with search.php?query=o'niel&action=results, then assign a block to that page. Finally, perform a search for o'niel). I doubt anyone has created a similar symlink, or we would have heard of the problem before now.