This addresses #767 - adsense tags will now work in custom blocks and other textareas that are displayed using icms_core_Textsanitizer::displayTarea. They will not work in custom blocks that are type 'HTML', because that content is not passed through displayTarea. This is also the reason customtags do not work in a custom block of type 'HTML'
When we can process input and output filtering separately, we will have a solution that will allow admins to create custom blocks with scripts, iframes and other restricted content, because we can filter that on input based on the user's role. Trying to filter on output based on user role would be inefficient and not a complete solution.