Making objects use dataFilter checkVar function for txtArea data_types.
Output should be fine, I'm not entirely sure if i have done it properly in the handler insert() function though, needs thoroughly testing.

if it works correctly anything using the IPF object handler with data_type of XOBJ_DTYPE_TXTAREA should now be filtered on input aswell as light filtering on output automatically, so no need to use checkvar in the modules.