Reverting changes from #11878
I removed the html comments yesterday as it was causing more trouble than it's worth. it worked as far as showing where input filtering is being done, & unfortunately it showed some flaws in the core. we still have a hell of a lot of work to do in order to achieve a proper core where we can filter on input properly.