Merging a bunch of changes from the 1.3 branch. Revisions 11612 - 11960
As 2.0 is diverging more from the 1.3 and older code, merging is getting more difficult. Once your changes in the 1.3 branch are tested by another developer/tester, please merge them to the trunk. Where you will run into challenges are changes made in the system module, since a lot of files are moved and/or renamed. Taking a little bit of time, merges can follow the moves and renames. What you'll need to resolve are 'tree conflicts', not just file conflicts.