Use reg users group if no default module/page set
Currently, if the Registered Users group has a default module/page set, it is ignored if the user is a member of another group that has no default module/page set.

Perhaps the Registered Users group should be used as a failsafe, so that you don't have to set default pages for every single group in your site? The only downside I can see is that perhaps webmasters would have specifically set a certain group to have no default page and intend that as the behaviour for that group. I think it's more powerful to rely on Registered Users to avoid people having to set a default page for every group though. Especially on sites with many many groups.