Updated string filters, though they are not used in the core. See #1422
Even though FILTER_UNSAFE_RAW isn't recommended on its own, this use in
combination with strip_tags() and the filter flags does produce safe
results that are almost identical to the filter ID that was in use
(FILTER_SANITIZE_STRING) and its flags. These options are not found in
use in any core files.