Hashing the reset password key for more security
With the hashing, the portions of the key will no longer be open to
discovery. I've also added another test so if the key is wrong, it
redirects to the user page and does not send another email. This way,
the user will not get flooded with emails if someone is attempting a
brute force attack on their account.

I'll leave the language constants as is for now.