Update language constants for users (#1544)
* Making sure password reset key is unique and temporary

Change in the structure of the field over time has made the old partial
string common for all users. This change makes it unique and temporary,
basing a portion of the key on their last login. Cleaned up the email
subjects a bit, too.

* Hashing the reset password key for more security

With the hashing, the portions of the key will no longer be open to
discovery. I've also added another test so if the key is wrong, it
redirects to the user page and does not send another email. This way,
the user will not get flooded with emails if someone is attempting a
brute force attack on their account.

* Update to language constants for site name (site URL) for consistency

Discovered while addressing the lost password user key

* some small corrections in FR and NL

---------

Co-authored-by: david <david@davidjanssens.be>