Fix: icms_core_Object cleanVars and getVar errors with empty arrays (#1623)
* Removing extra code block for users table fields

I'm not sure how this got duplicated

* Make sure an array is returned when retrieving an array See #1621

* Empty arrays were being skipped instead of serialized

I'm including this with #1621 - if an array was empty when cleanVars was
called, the array was skipped instead of serialized. This caused fatal
errors when trying to further process the variable. See further
explanation in https://www.impresscms.org/modules/iforum/viewtopic.php?post_id=51053#forumpost51053

* use exceptions to catch unserialize errors

Co-authored-by: David Janssens <david.j@impresscms.org>
Co-authored-by: Raimondas Rimkevičius <github@mekdrop.name>